refactor(hypervisor): share governed authority verification
